DB: Updated controllers for new requirements.
Rakefet required to add notes and project fields. controller methods for CognitiveTest were added accordingly, in order to filter search by these fields.
Also, the way we sent the tests to the client was very inefficient - we sent all tests with all blocks and questions in advance, when most of the tests are not needed for it.
This was changed by adding and changing the API of the controllers.
for example: findTestForTestManager was changed to return only a list of CognitiveTest (without the questions).

#160