DB: Removed fields lastModified and lastAnswered from DB.
These two fields, that belonged to CognitiveTest class, were fields of type - Date.
The type format used in the mySql DB was different than the format that was used in the code by Spring.
This caused a problem that made the field in the DB null, although it was not null in the code.
This complication led us to the understanding that we dont actually need these fields (not in HLD).
And so, they were removed.
Changes that occurred: removed fields from db_settings file, removed fields from classes in code, removed parts of tests that checked the value of these fields.
 Checked that all tests (apart from integration, since they are open in another issue) pass (including DAO's).

#173