Backend: Added Microsoft's Azure Application Insights Capabilities.
As a feature that needed to be supported, microsoft's service application insights was added to the project.
A config class was added, connecting to Azure with an instrumental key.
This includes adding code to track events, failures (exception) and http requests in the server side.
Any further tracking abilities can be added very easily via using a singleton class CognitivityApplicationInsights.
 Additions to the client side (the web app) should be inserted to the client JS code.
 Will open issue for that.
 Exactly which event we would like to track will be talked over in the next team meeting.

#178