Backend: Changed parameter of service (FileLoader) to string (as body of HTTP message).
The previous parameter was a file path. This was just for testing on my PC, and it obviously wouldnt have worked on the real server. So the change just swaps the reading of the file, to reading the json from a string (in json format) to be passed on from the client code in the HTTP POST message's body.
Need to remove loadFromDirectory feature - its no good.

#206